+++ test/supercluster_mutable_test.dart
@@ -381,4 +381,20 @@ void main() {
 
     expect(clusterDataPerLayer, equals(clusterDataPerLayerAfterModification2));
   });
+
+  test('contains', () {
+    final index = supercluster(Fixtures.features);
+    for (final feature in Fixtures.features) {
+      expect(index.contains(feature), isTrue);
+    }
+
+    var feature = Map<String, dynamic>.from(Fixtures.features.first);
+    feature['properties']['name'] = 'Changed name';
+    expect(index.contains(feature), isFalse);
+
+    feature = Map<String, dynamic>.from(Fixtures.features.first);
+    feature['geometry']['coordinates'][0] =
+        feature['geometry']['coordinates'][0] + 1;
+    expect(index.contains(feature), isFalse);
+  });
 }